Move /check to CardReaderController
