+++ app/Resources/views/default/card_reader/index.html.twig
@@ -39,7 +39,7 @@
                 e.preventDefault();
             });
         })
-        var barcode_submit_url = "{{ path('check') }}";
+        var barcode_submit_url = "{{ path('card_reader_check') }}";
     </script>
     <script src="{{ asset('bundles/app/js/barcode.js') }}"></script>
     <script type="text/javascript">
