+++ app/Resources/views/default/index.html.twig
@@ -208,7 +208,7 @@
         });
     })
     {% if not is_granted("IS_AUTHENTICATED_REMEMBERED") %}
-        var barcode_submit_url = "{{ path('check') }}";
+        var barcode_submit_url = "{{ path('card_reader_check') }}";
     {% endif %}
 </script>
 {% if not is_granted("IS_AUTHENTICATED_REMEMBERED") %}
