+++ app/Resources/views/user/check.html.twig
@@ -106,18 +106,18 @@
             </style>
         {% endif %}
     </div>
-    <a id="close" href="{{ path("cardReader") }}" class="black-text"><i class="material-icons large">close</i></a>
+    <a id="close" href="{{ path('cardReader') }}" class="black-text"><i class="material-icons large">close</i></a>
 
 {% endblock %}
 
 {% block javascripts %}
     <script>
         window.setTimeout(function(){
-            window.location.href = "{{ path("cardReader") }}";
+            window.location.href = "{{ path('cardReader') }}";
         }, 10000);//10sec
 
         //can scan another barcode
-        var barcode_submit_url="{{ path('check') }}";
+        var barcode_submit_url="{{ path('card_reader_check') }}";
     </script>
     <script src="{{ asset('bundles/app/js/barcode.js') }}"></script>
 {% endblock %}
