+++ src/AppBundle/Controller/DefaultController.php
@@ -9,11 +9,8 @@ use AppBundle\Entity\Membership;
 use AppBundle\Entity\Registration;
 use AppBundle\Entity\Shift;
 use AppBundle\Entity\ShiftBucket;
-use AppBundle\Entity\SwipeCard;
 use AppBundle\Entity\User;
 use AppBundle\Event\HelloassoEvent;
-use AppBundle\Event\SwipeCardEvent;
-use AppBundle\Event\ShiftValidatedEvent;
 use AppBundle\Form\AutocompleteBeneficiaryCollectionType;
 use AppBundle\Service\MembershipService;
 use AppBundle\Twig\Extension\AppExtension;
@@ -36,17 +33,6 @@ use Symfony\Component\Serializer\Encoder\JsonDecode;
 
 class DefaultController extends Controller
 {
-    /**
-     * @var boolean
-     */
-    private $swipeCardLogging;
-    private $swipeCardLoggingAnonymous;
-
-    public function __construct(string $swipeCardLogging, string $swipeCardLoggingAnonymous)
-    {
-        $this->swipeCardLogging = $swipeCardLogging;
-        $this->swipeCardLoggingAnonymous = $swipeCardLoggingAnonymous;
-    }
 
     /**
      * @Route("/", name="homepage")
@@ -208,55 +194,6 @@ class DefaultController extends Controller
             "buckets_upcoming" => $buckets_upcoming,
             "dynamicContent" => $dynamicContent
         ]);
-    }
-
-    /**
-     * @Route("/check", name="check", methods={"GET","POST"})
-     */
-    public function checkAction(Request $request)
-    {
-        $session = new Session();
-        $code = $request->get('swipe_code');
-        if (!$code) {
-            return $this->redirectToRoute('cardReader');
-        }
-        $em = $this->getDoctrine()->getManager();
-        if (!SwipeCard::checkEAN13($code)) {
-            return $this->redirectToRoute('cardReader');
-        }
-        $code = substr($code, 0, -1); //remove controle
-        $card = $em->getRepository('AppBundle:SwipeCard')->findOneBy(array('code' => $code, 'enable' => 1));
-        if (!$card) {
-            $session->getFlashBag()->add("error", "Oups, ce badge n'est pas actif ou n'existe pas");
-        } else {
-            $beneficiary = $card->getBeneficiary();
-            $membership = $beneficiary->getMembership();
-            $cycle_end = $this->get('membership_service')->getEndOfCycle($membership, 0);
-            $counter = $membership->getTimeCount($cycle_end);
-            if ($this->swipeCardLogging) {
-                $dispatcher = $this->get('event_dispatcher');
-                if ($this->swipeCardLoggingAnonymous) {
-                    $card = null;
-                }
-                $dispatcher->dispatch(SwipeCardEvent::SWIPE_CARD_SCANNED, new SwipeCardEvent($card, $counter));
-            }
-            $shifts = $em->getRepository('AppBundle:Shift')->getOnGoingShifts($beneficiary);
-            $dispatcher = $this->get('event_dispatcher');
-            foreach ($shifts as $shift) {
-                if ($shift->getWasCarriedOut() == 0) {
-                    $shift->validateShiftParticipation();
-                    $em->persist($shift);
-                    $em->flush();
-                    $dispatcher->dispatch(ShiftValidatedEvent::NAME, new ShiftValidatedEvent($shift));
-                }
-            }
-            return $this->render('user/check.html.twig', [
-                'beneficiary' => $beneficiary,
-                'counter' => $counter
-            ]);
-        }
-
-        return $this->redirectToRoute('cardReader');
     }
 
     /**
